Rename parsers to parser
