+++ sphinx_matlab/grammar/__init__.py
@@ -1,1 +1,1 @@
-from .parsers import GrammarParser
+from .parser import GrammarParser
